Integrate ladder filter, adjust osc/envelope parameters, and add button18 filter frequency mapping
- Include the new ladder.h header with a warning.
- Modify oscillator frequency multipliers and filter settings.
- Update envelope attack, decay, and release times.
- Add new volatile variables and button handling for filter frequency control.
- Remove extraneous Serial output in the sequencer.

+++ src/sequencer/SequencerDefs.h
@@ -37,13 +37,13 @@ struct Step {
   bool slide = false;     // Slide ON (true) or OFF (false)
   int note = 0;           // Note value, 0-24
   float velocity = 0.5f;  // Velocity, 0.0f - 1.0f
-  float filter = 0.0f;    // Filter value, 0.0f - 1.0f
+  int filter = 0;    // Filter value, 0.0f - 1.0f
 
   // Default constructor initializes to sensible defaults
   Step() = default;
 
   // Parameterized constructor for convenience
-  Step(bool g, bool s, int n, float v, float f)
+  Step(bool g, bool s, int n, float v, int f)
       : gate(g), slide(s), note(n), velocity(v), filter(f) {}
 };
